Update to merged code to support OS X with clang > 7 or gcc > 5.

+++ code-t1/tests/IO/FlashTest.cpp
@@ -90,7 +90,7 @@ TEST_GROUP(Flash)
 
     void simulateDeviceStatusWithRepeat(ioData status, int repeatCount)
     {
-        mock().expectNCalls(repeatCount, "IO_Read")
+        mock().expectNCalls((unsigned int) repeatCount, "IO_Read")
                 .withParameter("addr", StatusRegister)
                 .andReturnValue((int) status);
     }
